* Makefile.am: Re-add m4 directory * configure.ac: Use a real macro to generate stdint.h * libswfdec/actions.c: (pc_is_valid), (action_jump), (action_if), (action_constant_pool): Check PC values when jumping.  Fix memleak when freeing constant pools.  (From Eric Anholt) * libswfdec/proto.h: cleanup * libswfdec/swf.c: (tag_func_end), (tag_func_ignore): cleanup. Make Protect tag quieter. * libswfdec/swfdec_internal.h: include local stdint.h * libswfdec/swfdec_shape.c: (swf_shape_get_recs), (swf_shape_ignore_recs), (swf_morphshape_get_recs): fix gratuitous use of int16_t. * m4/ax_create_stdint_h.m4: add

+++ libswfdec/swfdec_shape.c
@@ -501,9 +501,9 @@ swf_shape_get_recs (SwfdecDecoder * s, SwfdecBits * bits, SwfdecShape * shape)
       edge_flag = swfdec_bits_getbits (bits, 1);
 
       if (edge_flag == 0) {
-        int16_t x0, y0;
-        int16_t x1, y1;
-        int16_t x2, y2;
+        int x0, y0;
+        int x1, y1;
+        int x2, y2;
 
         x0 = x;
         y0 = y;
@@ -957,9 +957,9 @@ swf_shape_ignore_recs (SwfdecDecoder * s, SwfdecBits * bits,
       edge_flag = swfdec_bits_getbits (bits, 1);
 
       if (edge_flag == 0) {
-        int16_t x0, y0;
-        int16_t x1, y1;
-        int16_t x2, y2;
+        int x0, y0;
+        int x1, y1;
+        int x2, y2;
 
         x0 = x;
         y0 = y;
@@ -1193,9 +1193,9 @@ swf_morphshape_get_recs (SwfdecDecoder * s, SwfdecBits * bits,
       edge_flag = swfdec_bits_getbits (bits, 1);
 
       if (edge_flag == 0) {
-        int16_t x0, y0;
-        int16_t x1, y1;
-        int16_t x2, y2;
+        int x0, y0;
+        int x1, y1;
+        int x2, y2;
 
         x0 = x;
         y0 = y;
